pyproject.toml
[project]
name = "exo-intel"
version = "0.1.0"
description = "A Django backend for exoplanet data, simulations and APIs."
authors = [
    {name = "Uriel Rodriguez",email = "codenameuriel28@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "amqp (==5.3.1)",
    "asgiref (==3.8.1)",
    "async-timeout (==5.0.1)",
    "billiard (==4.2.1)",
    "celery (==5.5.3)",
    "certifi (==2025.6.15)",
    "charset-normalizer (==3.4.2)",
    "click (==8.2.1)",
    "click-didyoumean (==0.3.1)",
    "click-plugins (==1.1.1.2)",
    "click-repl (==0.3.0)",
    "dj-database-url (==3.0.1)",
    "django (==5.2.3)",
    "django-filter (==25.1)",
    "djangorestframework (==3.16.0)",
    "graphene (==3.4.3)",
    "graphene-django (==3.2.3)",
    "graphql-core (==3.2.6)",
    "graphql-relay (==3.2.0)",
    "gunicorn (==23.0.0)",
    "idna (==3.10)",
    "kombu (==5.5.4)",
    "mypy-extensions (==1.1.0)",
    "packaging (==25.0)",
    "pathspec (==0.12.1)",
    "platformdirs (==4.3.8)",
    "promise (==2.3)",
    "prompt-toolkit (==3.0.51)",
    "psycopg2-binary (==2.9.10)",
    "python-dateutil (==2.9.0.post0)",
    "redis (==5.2.1)",
    "requests (==2.32.4)",
    "six (==1.17.0)",
    "sqlparse (==0.5.3)",
    "text-unidecode (==1.3)",
    "tomli (==2.2.1)",
    "typing-extensions (==4.14.0)",
    "tzdata (==2025.2)",
    "urllib3 (==2.5.0)",
    "vine (==5.1.0)",
    "wcwidth (==0.2.13)",
    "whitenoise (==6.9.0)",
    "django-environ (>=0.12.0,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
celery-types = "^0.23.0"
poethepoet = "^0.37.0"

[tool.poetry]
package-mode = false

[tool.poe]
# default
envfile = ".env.local"

[tool.poe.tasks."celery:start"]
cmd = "scripts/run_celery.sh"

[tool.poe.tasks."celery:stop"]
shell = "kill $(cat run/celery_worker.pid) $(cat run/celery_beat.pid) || true"

[tool.poe.tasks."celery:logs"]
shell = "tail -f logs/celery_worker.log logs/celery_beat.log"

[tool.poe.tasks]
runserver = "python3 manage.py runserver 0.0.0.0:7000"

makemigrations = "python3 manage.py makemigrations"

migrate = "python3 manage.py migrate"

createsuperuser = "python3 manage.py createsuperuser"

import_all_nasa_data = "python3 manage.py import_all_nasa_data"